+++ docs/changelog.md
@@ -4,6 +4,11 @@ All notable changes to Homeboy CLI are documented in this file.
 
 (This file is embedded into the CLI binary and is also viewable via `homeboy changelog`.)
 
+## Unreleased
+
+### Changed
+- Migrate component env detection to Extension API v1
+
 ## [0.377.1] - 2026-09-19
 
 ### Fixed